+++ Editor/ReadOnlyDrawer.cs
@@ -1,0 +1,26 @@
+﻿using Packages.UniKit.Runtime.Attributes;
+using UnityEditor;
+using UnityEngine;
+
+namespace Packages.UniKit.Editor
+{
+    [CustomPropertyDrawer(typeof(ReadOnlyFieldAttribute))]
+    public class ReadOnlyDrawer : PropertyDrawer
+    {
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            return EditorGUI.GetPropertyHeight(property, label, true);
+        }
+
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            var oldState = GUI.enabled;
+
+            GUI.enabled = false;
+
+            EditorGUI.PropertyField(position, property, label);
+
+            GUI.enabled = oldState;
+        }
+    }
+}